requirements.txt
beautifulsoup4==4.11.1
huggingface_hub==0.7.0
nltk==3.7
numpy==1.22.4
pandas==1.4.2
python-dotenv==0.20.0
ratelimit==2.2.1
requests==2.28.0
seqeval==1.2.2
tabulate==0.8.9
torch==1.11.0
tqdm==4.64.0
transformers==4.20.0
wandb==0.12.18
